requirements.txt
asgiref==3.11.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==6.0
django-cors-headers==4.9.0
django-fernet-fields==0.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
google-auth==2.45.0
idna==3.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
requests==2.32.5
rsa==4.9.1
sqlparse==0.5.5
urllib3==2.6.2
